Move `StartupOptions` to a separate file to reduce the size of `app.rs`
